[squash] pr comments
