+++ src/controllers/simplified-account/settings/service-name/service-name.controller.js
@@ -1,5 +1,20 @@
+const formatSimplifiedAccountPathsFor = require('@utils/simplified-account/format/format-simplified-account-paths-for')
+const paths = require('@root/paths')
+const { response } = require('@utils/response')
+
+function get (req, res) {
+  const context = {
+    messages: res.locals?.flash?.messages ?? [],
+    serviceNameEn: req.service.serviceName.en,
+    serviceNameCy: req.service.serviceName.cy,
+    manageEn: formatSimplifiedAccountPathsFor(paths.simplifiedAccount.settings.serviceName.edit, req.service.externalId, req.account.type),
+    manageCy: formatSimplifiedAccountPathsFor(paths.simplifiedAccount.settings.serviceName.edit, req.service.externalId, req.account.type) + '?cy=true'
+  }
+  return response(req, res, 'simplified-account/settings/service-name/index', context)
+}
+
 module.exports = {
-  index: require('./index/service-name-index.controller'),
+  get,
   edit: require('./edit/edit-service-name.controller'),
   removeWelshServiceName: require('./remove-cy/remove-welsh-service-name.controller')
 }
